pyproject.toml
[tool.poetry]
name = "polar-bites"
version = "0.1.0"
description = "Extra bites for py-polars"
authors = ["maxastyler <maxastyler@gmail.com>"]
readme = "README.md"
packages = [{ include = "polar_bites" }]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
polars = "^0.20.6"
numpy = "^1.25.0"
scipy = "^1.12.0"
attrs = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pyright = "^1.1.304"
black = "^23.3.0"
debugpy = "^1.6.7"
hypothesis = "^6.72.1"
coverage = "^7.2.3"
sphinx = "^7.0.1"
livereload = "^2.6.3"
sphinx-autoapi = "^2.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
